feat: add option slot to radio component
